refactor: applied clean architecture for the project
--BREAKING CHANGE--
프로젝트의 패키지 구조를 클린 아키텍처에 맞춰 개편하였습니다.
