+++ app/src/androidTest/java/ksc/campus/tech/kakao/map/fake_module/FakeRepositoryModules.kt
@@ -4,12 +4,12 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.components.SingletonComponent
 import dagger.hilt.testing.TestInstallIn
-import ksc.campus.tech.kakao.map.models.repositories.MapRepositoryModule
-import ksc.campus.tech.kakao.map.models.repositories.MapViewRepository
-import ksc.campus.tech.kakao.map.models.repositories.SearchKeywordRepository
-import ksc.campus.tech.kakao.map.models.repositories.SearchKeywordRepositoryModule
-import ksc.campus.tech.kakao.map.models.repositories.SearchResultRepository
-import ksc.campus.tech.kakao.map.models.repositories.SearchResultRepositoryModule
+import ksc.campus.tech.kakao.map.domain.repositories.MapRepositoryModule
+import ksc.campus.tech.kakao.map.domain.repositories.MapViewRepository
+import ksc.campus.tech.kakao.map.domain.repositories.SearchKeywordRepository
+import ksc.campus.tech.kakao.map.domain.repositories.SearchKeywordRepositoryModule
+import ksc.campus.tech.kakao.map.domain.repositories.SearchResultRepository
+import ksc.campus.tech.kakao.map.domain.repositories.SearchResultRepositoryModule
 import ksc.campus.tech.kakao.map.repositoriesImpl.FakeMapViewRepository
 import ksc.campus.tech.kakao.map.repositoriesImpl.FakeSearchKeywordRepository
 import ksc.campus.tech.kakao.map.repositoriesImpl.FakeSearchResultRepository
