+++ app/src/main/java/ksc/campus/tech/kakao/map/domain/models/LocationInfo.kt
@@ -1,0 +1,8 @@
+package ksc.campus.tech.kakao.map.domain.models
+
+data class LocationInfo(
+    val address: String,
+    val name: String,
+    val latitude: Double,
+    val longitude: Double
+)
