+++ app/src/main/java/ksc/campus/tech/kakao/map/presentation/views/adapters/SearchKeywordClickCallback.kt
@@ -1,0 +1,6 @@
+package ksc.campus.tech.kakao.map.presentation.views.adapters
+
+interface SearchKeywordClickCallback {
+    fun clickKeyword(keyword: String)
+    fun clickRemove(keyword: String)
+}
