--- app/src/main/java/ksc/campus/tech/kakao/map/views/adapters/SearchKeywordClickCallback.kt
@@ -1,6 +1,0 @@
-package ksc.campus.tech.kakao.map.views.adapters
-
-interface SearchKeywordClickCallback {
-    fun clickKeyword(keyword: String)
-    fun clickRemove(keyword: String)
-}
