+++ app/src/main/res/layout/fragment_kakao_map.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".views.fragments.KakaoMapFragment">
+    tools:context=".presentation.views.fragments.KakaoMapFragment">
 
     <com.kakao.vectormap.MapView
         app:layout_constraintRight_toRightOf="parent"
