+++ app/src/main/res/layout/fragment_search_result.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ksc.campus.tech.kakao.map.views.fragments.SearchResultFragment">
+    tools:context="ksc.campus.tech.kakao.map.presentation.views.fragments.SearchResultFragment">
 
     <TextView
         android:id="@+id/text_no_result"
