+++ app/src/main/res/layout/search_info_bottomsheet.xml
@@ -4,6 +4,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".views.MainActivity">
+    tools:context=".presentation.views.MainActivity">
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
